requirements.txt
beautifulsoup4>=4.12
lxml>=5.0
openpyxl>=3.1
pillow>=10.0
